a dirty workaround
There have been some initial launch problems, mostly because of lack of SSL encryption on the backend server (you can't get SSL certs for IP adresses; a domain is required). Thus, the decision was made to host the entire event on the backend server only. This is just a quick workaround so the old links will still work.